// frontend/src/api/models/SchemaUserSegmentVideo.ts
export interface SchemaUserSegmentVideo {
    /**
     * 
     * @type {SchemaUserVideoDetails}
     * @memberof SchemaUserSegmentVideo
     */
    segmentDetails: SchemaUserVideoDetails;
    /**
     * 
     * @type {string}
     * @memberof SchemaUserSegmentVideo
     */
    parentVideoId: string;
    /**
     * 
     * @type {string}
     * @memberof SchemaUserSegmentVideo
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof SchemaUserSegmentVideo
     */
    videoDuration: number | null;
    /**
     * 
     * @type {VideoAspectRatio}
     * @memberof SchemaUserSegmentVideo
     */
    videoAspectRatio: VideoAspectRatio;
    /**
     * 
     * @type {VideoType}
     * @memberof SchemaUserSegmentVideo
     */
    videoType: VideoType;
    /**
     * 
     * @type {VideoFormat}
     * @memberof SchemaUserSegmentVideo
     */
    videoFormat: VideoFormat;
    /**
     * 
     * @type {string}
     * @memberof SchemaUserSegmentVideo
     */
    videoResolution: string;
    /**
     * 
     * @type {DissaminateOnSocialMediaPlatforms}
     * @memberof SchemaUserSegmentVideo
     */
    dissaminateOnSocialMediaPlatforms: DissaminateOnSocialMediaPlatforms;
    /**
     * 
     * @type {string}
     * @memberof SchemaUserSegmentVideo
     */
    shortDescription: string;
    /**
     * 
     * @type {string}
     * @memberof SchemaUserSegmentVideo
     */
    longDescription: string;
    /**
     * 
     * @type {Tags}
     * @memberof SchemaUserSegmentVideo
     */
    tags: Tags;
    /**
     * 
     * @type {string}
     * @memberof SchemaUserSegmentVideo
     */
    thumbnailUrl?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SchemaUserSegmentVideo
     */
    videoUrl?: string | null;
    /**
     * 
     * @type {CreatedAt}
     * @memberof SchemaUserSegmentVideo
     */
    createdAt: CreatedAt;
    /**
     * 
     * @type {UpdatedAt}
     * @memberof SchemaUserSegmentVideo
     */
    updatedAt: UpdatedAt;
}